% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bangkok1}
\alias{bangkok1}
\title{Annual Maximum Daily Rainfall for Bangkok, Thailand}
\format{
A data frame with 75 rows and 2 variables:
\describe{
\item{year}{The observation year (1951-2025).}
\item{rainfall}{The annual maximum daily rainfall in millimeters (mm).}
}
}
\source{
Thai Meteorological Department (TMD)
}
\usage{
data(bangkok1)
}
\description{
A dataset containing the annual maximum series (AMS) of daily rainfall
measurements in Bangkok, Thailand, from 1951 to 2025. This dataset is
typically used for extreme value analysis and hydrological frequency modeling.
}
\examples{
data(bangkok1)
head(bangkok1)
}
\keyword{datasets}
