% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.parpe3.R
\name{cal_Lam}
\alias{cal_Lam}
\title{Calculate the Lambda Function for the PE3 Distribution}
\usage{
cal_Lam(alpha = NULL, eta = 1, r = 1)
}
\arguments{
\item{alpha}{A numeric shape parameter.}

\item{eta}{A non-negative integer representing the order of the LH-moments. Default is 1.}

\item{r}{An integer representing the moment order index. Default is 1.}
}
\value{
A numeric value representing the Lambda component.
}
\description{
An internal helper function to compute the weighted sum of the Omega function
values, calculating the specific Lambda components for the PE3 parameter estimation.
}
\keyword{internal}
