% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{khonkaen}
\alias{khonkaen}
\title{Annual Maximum Daily Rainfall for Khon Kaen, Thailand}
\format{
A data frame with 42 rows and 2 variables:
\describe{
\item{year}{The observation year (1984-2025).}
\item{rainfall}{The annual maximum daily rainfall in millimeters (mm).}
}
}
\source{
Thai Meteorological Department (TMD)
}
\usage{
data(khonkaen)
}
\description{
A dataset containing the annual maximum series (AMS) of daily rainfall
measurements in Khon Kaen, Thailand, from 1984 to 2025.
}
\examples{
data(khonkaen)
head(khonkaen)
}
\keyword{datasets}
