% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHmom-package.R
\docType{package}
\name{LHmom-package}
\alias{LHmom}
\alias{LHmom-package}
\title{LHmom: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
