% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.pargev.R
\name{lhmom.gev}
\alias{lhmom.gev}
\title{Calculate Theoretical LH-moments for the Generalized Extreme Value (GEV) Distribution}
\usage{
lhmom.gev(para, eta = NULL)
}
\arguments{
\item{para}{A numeric vector of three parameters: c(xi, alpha, k), corresponding
to the location, scale, and shape parameters of the GEV distribution, respectively.}

\item{eta}{A non-negative integer (between 0 and 4) representing the order
of the LH-moments.}
}
\value{
A list containing:
\itemize{
\item \code{lambdas}: A named numeric vector of the first four theoretical LH-moments.
\item \code{ratios}: A named numeric vector of the corresponding LH-moment ratios.
\item \code{eta}: The order of the LH-moments used.
}
}
\description{
This function computes the theoretical LH-moments and LH-moment ratios for
the Generalized Extreme Value (GEV) distribution given its parameters.
The computations are derived based on the formulas provided by Wang (1997).
}
\references{
Wang, Q. J. (1997). Using higher order L-moments for regional
flood frequency analysis. \emph{Water Resources Research}, 33(12), 2841-2848.
}
