% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo.lhmom.others.R
\name{lhmom.ggd}
\alias{lhmom.ggd}
\title{Theoretical LH-moments for the Generalized Gumbel (GGD) Distribution}
\usage{
lhmom.ggd(para = NULL, eta = 1)
}
\arguments{
\item{para}{A numeric vector of three parameters: c(mu, sigma, h).}

\item{eta}{The order of LH-moments (default is 1).}
}
\value{
A list containing the calculated LH-moments and the distribution type ("ggd").
}
\description{
This function computes the theoretical LH-moments for the Generalized
Gumbel (GGD) distribution. It is evaluated as a special case of the
four-parameter Kappa distribution with the shape parameter k set to 0.
}
