% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.pargno.R
\name{lhmom.gno}
\alias{lhmom.gno}
\title{Theoretical LH-moments function of GNO distribution}
\usage{
lhmom.gno(para = NULL, eta = 1)
}
\description{
Function to calculate theoretical LH-moments for GNO distribution
}
