% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo.lhmom.others.R
\name{lhmom.gum}
\alias{lhmom.gum}
\title{Theoretical LH-moments for Gumbel Distribution}
\usage{
lhmom.gum(para = NULL, eta = 1)
}
\arguments{
\item{para}{A vector of parameters c(mu, sigma)}

\item{eta}{The order of LH-moments (default is 1).}
}
\value{
A list containing the LH-moments and distribution type.
}
\description{
This function computes the theoretical LH-moments for the Gumbel distribution
by evaluating it as a special case of the Generalized Extreme Value (GEV)
distribution with the shape parameter set to zero.
}
