% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo.lhmom.others.R
\name{lhmom.k3d.kfix}
\alias{lhmom.k3d.kfix}
\title{Theoretical LH-moments for the Three-Parameter Kappa Distribution with Fixed k}
\usage{
lhmom.k3d.kfix(para = NULL, eta = 1, kfix = 0)
}
\arguments{
\item{para}{A numeric vector of three parameters: c(mu, sigma, h).}

\item{eta}{The order of LH-moments (default is 1).}

\item{kfix}{The fixed numeric value for the shape parameter k (default is 0).}
}
\value{
A list containing the calculated LH-moments and the distribution type ("kap").
}
\description{
This function computes the theoretical LH-moments for the three-parameter
Kappa distribution where the shape parameter k is fixed.
}
