% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhmoms.R
\name{lhmoms}
\alias{lhmoms}
\title{Calculate Sample LH-moments}
\usage{
lhmoms(data, eta = NULL, nmom = 5)
}
\arguments{
\item{data}{A numeric vector of data values.}

\item{eta}{A non-negative integer (or a sequence of integers) between 0 and 4
representing the order of the LH-moments.}

\item{nmom}{An integer specifying the maximum number of moments to compute (default is 5).}
}
\value{
A list containing:
\itemize{
\item \code{eta}: The order(s) of the LH-moments.
\item \code{lambdas}: A matrix of the calculated sample LH-moments.
\item \code{ratios}: A matrix of the calculated sample LH-moment ratios.
}
}
\description{
This function computes the sample LH-moments and LH-moment ratios from a numeric
dataset. The estimation is based on the methodology proposed by Wang (1997).
The order parameter \code{eta} determines the weight assigned to larger observations.
When \code{eta = 0}, the calculation reduces to the ordinary sample L-moments.
}
\references{
Wang, Q. J. (1997). Using higher order L-moments for regional
flood frequency analysis. \emph{Water Resources Research}, 33(12), 2841-2848.
}
