% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{temp_khonkaen}
\alias{temp_khonkaen}
\title{Annual Maximum Daily Temperature for Khon Kaen}
\format{
A data frame with 42 rows and 2 variables:
\describe{
\item{year}{The observation year (1985-2025).}
\item{max_temp}{The annual maximum daily temperature in degrees Celsius (°C).}
}
}
\source{
Thai Meteorological Department (TMD)
}
\usage{
data(temp_khonkaen)
}
\description{
A dataset containing the annual maximum series (AMS) of daily temperature
measurements in Khon Kaen, Thailand, from 1985 to 2025. This dataset is
useful for modeling extreme heat events.
}
