% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{temp_mahasarakham}
\alias{temp_mahasarakham}
\title{Annual Maximum Daily Temperature for Maha Sarakham}
\format{
A data frame with 41 rows and 2 variables:
\describe{
\item{year}{The observation year (1985-2025).}
\item{Tmax}{The annual maximum daily temperature in degrees Celsius (°C).}
}
}
\source{
Thai Meteorological Department (TMD)
}
\usage{
data(temp_mahasarakham)
}
\description{
A dataset containing the annual maximum series (AMS) of daily temperature
measurements in Maha Sarakham, Thailand, from 1985 to 2025.
}
