% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wang.test.lhgev.R
\name{wang.test.lhgev}
\alias{wang.test.lhgev}
\title{Wang's Goodness-of-Fit Test for the Generalized Extreme Value (GEV) Distribution}
\usage{
wang.test.lhgev(data)
}
\arguments{
\item{data}{A numeric vector of data values.}
}
\value{
A data frame containing the GOF test results for each \code{eta} value
(from 0 to 4), with the following columns:
\itemize{
\item \code{eta}: The order of the LH-moments.
\item \code{z.test}: The calculated Z-statistic for the GOF test.
\item \code{cond.sigma}: The conditional standard deviation of the sample LH-kurtosis.
\item \code{p.value}: The two-sided p-value corresponding to the Z-test statistic.
}
}
\description{
This function performs a Goodness-of-Fit (GOF) test for the Generalized Extreme
Value (GEV) distribution using LH-moments, based on the methodology proposed
by Wang (1998). It calculates a Z-test statistic by comparing the sample
LH-kurtosis with the theoretical LH-kurtosis. The test is evaluated across
LH-moment orders (\code{eta}) from 0 to 4.
}
\references{
Wang, Q. J. (1998). Approximate goodness-of-fit tests of fitted
generalized extreme value distributions using LH moments. \emph{Water Resources
Research}, 34(12), 3497-3502.
}
