% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.pargpa_glo.R
\name{lh.parglo}
\alias{lh.parglo}
\title{Estimate Parameters of the Generalized Logistic (GLO) Distribution using LH-moments}
\usage{
lh.parglo(data, eta = 1)
}
\arguments{
\item{data}{A numeric vector of data values.}

\item{eta}{A non-negative integer (between 0 and 4) representing the order
of the LH-moments. Default is 1.}
}
\value{
A list containing:
\itemize{
\item \code{type}: The distribution type (\code{"glo"}).
\item \code{para}: A named numeric vector containing the estimated parameters
(\code{mu} for location, \code{sigma} for scale, \code{k} for shape).
\item \code{eta}: The order of the LH-moments used.
\item \code{source}: The name of the function (\code{"lh.parglo"}).
\item \code{ifail}: A numeric indicator of success (0 for success).
}
}
\description{
This function estimates the parameters of the Generalized Logistic (GLO)
distribution based on the sample LH-moments. The analytical formulas for
the parameter estimation are derived based on the methodology presented
by Meshgi and Khalili (2009).
}
\references{
Meshgi, A., & Khalili, D. (2009). Comprehensive evaluation of
regional flood frequency analysis by L- and LH-moments. \emph{Stochastic
Environmental Research and Risk Assessment (SERRA)}, 23(1), 137-152.
}
