% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lh.pargno.R
\name{lh.pargno}
\alias{lh.pargno}
\title{LH-moments estimation parameters for GNO distribution}
\usage{
lh.pargno(data, eta = NULL, opt = FALSE, ntry = 5)
}
\description{
Function to calculate LH-moments estimation parameters for GNO distribution
}
