% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHmom-package.R
\docType{package}
\name{LHmom-package}
\alias{LHmom}
\alias{LHmom-package}
\title{LHmom: Parameter Estimation and Total Calculations of Linear Higher-Order Moments (LH-Moments)}
\description{
Computes sample and theoretical Linear Higher-Order Moments (LH-moments) and implements parameter estimation for various statistical distributions, including the Generalized Extreme Value (GEV), Generalized Logistic (GLO), Generalized Pareto (GPA), Generalized Normal (GNO), Pearson Type III (PE3), and Kappa distributions. To support robust extreme value analysis, the package provides essential analytical tools, such as visual diagnostic capabilities via universal Quantile-Quantile plots and Wang's goodness-of-fit test. The design, parameterization, and output structures of this package closely follow the style and conventions of the 'lmomco' package to ensure a familiar and consistent interface for users accustomed to L-moment calculations in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://palakorn-seenoi.github.io/LHmom/}
}

}
\author{
\strong{Maintainer}: Palakorn Seenoi \email{palakorns@kku.ac.th}

Authors:
\itemize{
  \item Yire Shin \email{shinyire@daum.net}
  \item Piyapatr Busababodhin \email{piyapatr.b@msu.ac.th}
  \item Jeong-Soo Park \email{jspark@jnu.ac.kr}
}

}
\keyword{internal}
